move routes.config to config.js
