+++ src/config.js
@@ -89,3 +89,20 @@ export const shortTypeMapping = {
   'kedro.extras.datasets.tracking.json_dataset.JSONDataSet': 'tracking',
   'kedro.extras.datasets.tracking.metrics_dataset.MetricsDataSet': 'tracking',
 };
+
+// URL parameters for each element/section
+export const params = {
+  focused: 'focused_id=',
+  selected: 'selected_id=',
+  pipeline: 'pipeline_id=',
+};
+
+const activePipeline = `${params.pipeline}:pipelineId`;
+
+export const routes = {
+  flowchart: {
+    main: '/',
+    focusedNode: `/?${activePipeline}&${params.focused}:id`,
+    selectedNode: `/?${activePipeline}&${params.selected}:id`,
+  },
+};
